+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/DOTS/EntityInstantiator.cs
@@ -175,7 +175,7 @@ namespace GLTFast
             MeshResult meshResult,
             uint[] joints = null,
             uint? rootJoint = null,
-            float[] morphTargetWeights = null,
+            IReadOnlyList<float> morphTargetWeights = null,
             int meshNumeration = 0
         )
         {